media: mfc: change nal_q_exception log for multiframe
Even if there is a multiframe, it can go back to NALQ mode.
This causes the NALQ log to be shown whenever the multiframe is
detected. This is modified so that the NALQ log is shown only when
the first multiframe detect.

Bug: 287512217
Test: Youtube Playback
Signed-off-by: Sunyoung Kang <sy0816.kang@samsung.com>
Change-Id: I6d0ba1e0157ae50452a8bd4c9d797052520e62fc
(cherry picked from commit 30a27b95894c1953032e4e9ec614dfc2747f7f81)
Signed-off-by: Will McVicker <willmcvicker@google.com>