pkvm-s2mpu: Control Zuma S2MPUs from the kernel module
The android13-5.15 GKI kernel used by Zuma does not support S2MPUv9 and
the corresponding changes cannot be merged. Control the S2MPUs from the
kernel module, as opposed to the hypervisor, until the base kernel is
updated to android14-x.

The kernel module enables the S2MPUs for all VIDs and configures the
FMPTs to PROT_RW on resume and PROT_NONE on suspend.

Bug: 240228255
Signed-off-by: David Brazdil <dbrazdil@google.com>
Change-Id: Iee392f1b155b9964be46c101febd881de8c3e43b
Signed-off-by: Will McVicker <willmcvicker@google.com>