arm64/s2mpu: Register driver with pKVM explicitly
Add a register function to the S2MPU driver to register the
driver with pKVM.

Previously, this was done implicitly on the first device register.
The counterpart of this change is done in GKI in aosp/2303645.

This change was done because the init function requires s2mpu version
passed as an argument. so we call it explictily instead of passing the
version to register function.

Change-Id: I69c6cf31027af7fb7d38c436b30e594edeca9f85
Signed-off-by: Mostafa Saleh <smostafa@google.com>
Signed-off-by: Will McVicker <willmcvicker@google.com>