arm64/s2mpu: Add module param for printing caches.
On S2MPU faults, it will print different caches (MPTC, PTLB, STLB)
which can print 20K lines for one faults. Most of the time the useful
information is the address of the fault and direction which can be
lost in the log(last_kmsg) if many caches are printed.
So a new module parameter(print_caches) is added that can be set from
the kernel cmdline to control the behavior of printing caches at fault.

Change-Id: Iccee552441347585a384a242e56392c6862a9d78
Signed-off-by: Mostafa Saleh <smostafa@google.com>
Signed-off-by: Will McVicker <willmcvicker@google.com>