arm64/s2mpu: Force finalize S2MPU after probing all devices.
Previously, for each device probed, it will check if it is the last
one it will call finalize.

However, this doesn't guarantee that the all devices are probed, and
this would allow IOMMU driver/device registration to be done after
boot is complete, if one or more devices are not probed.

So, the driver is changed to probe all devices in driver init, and
after the init, finalize would be called. If not all devices are
probed it will issue a Warning.

Also, remove atomic from nr_devs_registered as probed is not done in
parallel.

Change-Id: I0aa586df1c7d69da438a2ff889af42da34102e10
Signed-off-by: Mostafa Saleh <smostafa@google.com>
Signed-off-by: Will McVicker <willmcvicker@google.com>