arm64/s2mpu: Enable runtime PM for all S2MPUs
Previously, we enabled runtime PM only for those S2MPU instances that
were linked to a power domain or had the device tree property
"always-on" set.

With this change, we want to support a third type of configuration where
the S2MPU is neither always on nor associated with a power domain. In
that configuration, runtime PM will now still be enabled which allows a
client driver to runtime PM control the S2MPU via calls to
pm_runtime_get_sync and related functions.

The specific use case we want to enable is GSA which uses runtime PM not
because its S2MPU is actually powered down but to make pKVM pause
register reads and writes during low power states in which such
read/write accesses would be extremely slow due to its clock speed being
reduced by a factor of 16.

Bug: 265407176
Change-Id: I29bd942848ad320ae0bde7e223fe2b4477ae1d6b
Signed-off-by: Mostafa Saleh <smostafa@google.com>
Signed-off-by: Daniel Mentz <danielmentz@google.com>
Signed-off-by: Will McVicker <willmcvicker@google.com>