arm64/s2mpu: Use new pkvm_iommu_register function signature
pkvm_iommu_register was changed in ACK to accept an extra argument
"flags", so device specfic flags can be passed to the hypervisor
driver as presence of sync devices.
This patch changes pkvm_iommu_register calls to have the new arg.

The flags will be used later when sync devices are supported.

Bug: 257544191
Change-Id: Id2d6371d6d688aeb07e7da6a879a1bd9348a4233
Signed-off-by: Mostafa Saleh <smostafa@google.com>
Signed-off-by: Will McVicker <willmcvicker@google.com>