arm64/dts: gs201: Early load s2mpu module and dependencies
Recently, early loading was introduced for pKVM modules. It is enabled
by the same cmdline option but with a list of modules as an argument. In
that case, loading is done by the kernel itself before host deprivilege,
thereby reducing the attack surface.

This is a port from ZUMA change:
commit 96f599bca292 ("arm64/dts: Early pKVM module loading of exynos-pd,pkvm_s2mpu-v9")

Change-Id: I11248244aef0292cf8df2b29b4d6fbf0a00dfaa6
Signed-off-by: Mostafa Saleh <smostafa@google.com>
Signed-off-by: Will McVicker <willmcvicker@google.com>