arm64/s2mpu: Relax device probe if pKVM disabled
We used to force probe all device at init when pKVM is present, as we
need to know that DMA protection is in place before the we de-privilege
the device. However, if pKVM is not there we don't need to do that,
which is useful because we can now support deferred probing.

Change-Id: I7e1d4e0cfbab5de6b8f41b2d0ff01df86eb088ab
Signed-off-by: Mostafa Saleh <smostafa@google.com>
Signed-off-by: Will McVicker <willmcvicker@google.com>