misc: logbuffer: Add a module parameter for the number of buffer entries
In some situation, logbuffer size (default 1024 entries) may not be
enough and the key logs may rolled over. For the debugging purpose, add
a new module parameter so that a larger entry number can be passed to
logbuffer through kernel command line. This parameter is an unsigned
integer as a power of two which is then calculated to the number of
logbuffer entries. If the parameter is larger than 20, it will be
limited in 20 to avoid large memory usage. If it is less than 3, it will
be set to 3.

Also do some tweaks such as:
- Change the type of _head / _tail variables to uint
- Change the overflow detection of the ring buffer from "MOD the entry
  number" to "AND the entry mask" to avoid the division operation
- Rename the constants to LOGBUFFER_ prefix
- Remove some redundant code

Bug: 295474320
Signed-off-by: Kyle Tso <kyletso@google.com>
(ported from commit c037e49e64aa5e31f6ed06238d3e735f7faf3d86)
Merged-In: If90207c0c03dda87bace9ab45d4a083ddd702d91
Change-Id: If90207c0c03dda87bace9ab45d4a083ddd702d91
Signed-off-by: Will McVicker <willmcvicker@google.com>