tcpci_max77759: spin on chip->tcpci instead of using a lock
Commit "tcpci_max77759: don't start toggling when tcpci isn't present"
caused a lockdep report, preventing the use of lockdep to detect other
locking issues in the kernel. Because the crash that it was fixing
is rare, fix the original bug by spinning on chip->tcpci instead of
using a lock.

Bug: 289535202
Bug: 299701914
Signed-off-by: Peter Collingbourne <pcc@google.com>
Signed-off-by: RD Babiera <rdbabiera@google.com>
Change-Id: I8ca472176372bedc241524623417475618a8137a
Signed-off-by: Will McVicker <willmcvicker@google.com>