pcie: exynos: Remove pcie phy reset on resume
There is an issue on whi/pro that it may get stuck while resetting PCIe
PHY on resume. On zuma the step for resetting PCIe PHY on resume was
removed. Consulted with PHY team and confirmed that this change can be
applied to whi/pro as well.

Bug: 291708247
Change-Id: Ic0f50066c286ea3a34d5c16d2312dd594e31ff4b
Signed-off-by: tinghsin <tinghsin@google.com>
(cherry picked from commit a833adaca14823075bff57ca8c51d04cfa094590)
Signed-off-by: Will McVicker <willmcvicker@google.com>