gvotable: replace strlcpy() with strscpy() (v6.8 compatibility)
Linux commit [1] has removed strlcpy() and this causes this OOT pixel
driver to fail compilation:

    .../drivers/misc/gvotable.c: In function ‘gvotable_internal_update_reason’:
    .../drivers/misc/gvotable.c:245:9: error: implicit declaration of function ‘strlcpy’; did you mean ‘strscpy’? [-Werror=implicit-function-declaration]
      245 |         strlcpy(el->reason, new_reason, GVOTABLE_MAX_REASON_LEN);
          |         ^~~~~~~
          |         strscpy

strlcpy() reads the entire source buffer first. This read may exceed
the destination size limit. This is both inefficient and can lead
to linear read overflows if a source string is not NUL-terminated[2].
Additionally, it returns the size of the source string, not the
resulting size of the destination string. In an effort to remove
strlcpy() completely[3], replace strlcpy() here with strscpy().

Most locations here don't check the return value here, so a direct
replacement with strspy() is possible.

In the (unlikely?) event that gvotable_get_current_reason_unlocked()
wasn't able to copy the reason to the destination buffer fully, it will
now correctly report truncation errors (as -E2BIG):
    * There is one user in here that already deals with negative return
      values, gvotable_dump_election(), which is used for debugfs
      entries only. This will now report that in the debugfs output.
    * Currently, there are two users of gvotable_get_current_reason()
      in the Pixel OOT driver stack: bms/pmic-voter-compat.c and
      bms/p9221_charger.c.
      The pmic driver deals with negative return values, and percolates
      this up the call-chain, so those should be OK.
      bms/p9221_charger.c doesn't deal with this, but a) the reason
      string appears to be used for (logbuffer) debug only, and b) the
      reason is still copied (albeit truncated).

While at it, use sizeof() for determining the maximum string legth
instad of open-coding.

Link: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=d26270061ae6 [1]
Link: https://www.kernel.org/doc/html/latest/process/deprecated.html#strlcpy [2]
Link: https://github.com/KSPP/linux/issues/89 [3]
Test: TH
Bug: 292499021
Change-Id: I6098f27ae372a84bfda21f153a9ea84947d32fc1
Signed-off-by: André Draszik <andre.draszik@linaro.org>