pixel_em: make sure all perf domains are initialized before probe
Since we didn't initialize the array `pixel_cluster_start_cpu`, we were
only checking if the perf domain for CPU 0 was initialized and not all
of the CPUs. This results in a random crash on boot when the start CPU
isn't CPU 0 and we try to finish probing before the start CPU perf
domain is initialized.

To make this issue easier to detect, add a BUG_ON when the perf domain
is NULL since it theoretically shouldn't ever be null when generating
the default EM profile.

Fixes: ca59be7b95b3 ("pixel_em: update to not depend on VH_SCHED")
Bug: 325158513
Change-Id: Iaf510af2fab2b41bd1c3498677449e0b8ca66e6c
Signed-off-by: Will McVicker <willmcvicker@google.com>