gs101: gsa: replace strlcpy() with strscpy() (v6.8 compatibility)
Linux commit [1] has removed strlcpy() and this causes this OOT pixel
driver to fail compilation.

strlcpy() reads the entire source buffer first. This read may exceed
the destination size limit. This is both inefficient and can lead
to linear read overflows if a source string is not NUL-terminated[2].
Additionally, it returns the size of the source string, not the
resulting size of the destination string. In an effort to remove
strlcpy() completely[3], replace strlcpy() here with strscpy().

Nothing here checks the return value here, so a direct replacement with
strspy() is possible.

Link: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=d26270061ae6 [1]
Link: https://www.kernel.org/doc/html/latest/process/deprecated.html#strlcpy [2]
Link: https://github.com/KSPP/linux/issues/89 [3]
Test: TH
Bug: 292499021
Change-Id: Ia3c483447a0ecf614bfeaf7c1fdb67a1cd78a01b
Signed-off-by: André Draszik <draszik@google.com>