dma-heap: samsung: rename MAX_ORDER to MAX_PAGE_ORDER (v6.8 compatibility)
Linux commit [1] has renamed MAX_ORDER to MAX_PAGE_ORDER. This was done
to highlight the fact that MAX_ORDER had been redefined with different
semantics in [2] and users might have needed adjustment.

This causes this OOT Pixel driver to fail compilation:

    .../soc/gs/drivers/dma-buf/heaps/samsung/samsung_heap.c:168:52: error: use of undeclared identifier 'MAX_ORDER'
      168 |         order = min_t(unsigned int, get_order(alignment), MAX_ORDER);
          |                                                           ^
    1 error generated.

Looking at the history:
    * MAX_ORDER used to be non-inclusive
    * it was changed to be inclusive in commit [2]
    * it was renamed (keeping new semantics) in [1]

The conversion here is straight-forward as the semantic change fixed the
incorrect use of MAX_ORDER here and the rename therefore doesn't affect
the code here. If anything, the code should have been using MAX_ORDER-1
before the semantic change in the first place because the maximum
allocation order the buddy allocator can deliver was MAX_ORDER-1.

[1] https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=5e0a760b4441
    ("mm, treewide: rename MAX_ORDER to MAX_PAGE_ORDER")
[2] https://lore.kernel.org/all/20230315113133.11326-1-kirill.shutemov@linux.intel.com/

Bug: 330466093
Test: TH
Change-Id: Iafe02426807ec0c55a968599a00d3d6774423561
Signed-off-by: André Draszik <draszik@google.com>