usb: dwc3: add a sysfs attribute to read data role
Add a read-only sysfs attribute new_data_role to present the data role
configured by typec stack. The attribute supports epoll so that it
can notify userspace.
Note that this attribute is different from the existing
dwc3_exynos_otg_state, which is the current running data role and is
only updated after the data role change has been completed. On the
contrary, new_data_role is updated before the driver attempts to change
the role.

Bug: 297224564
Test: read the sysfs attribute and verify epoll is working
Change-Id: I4136b0a42e7fba64a40f8e6f5b52571ad8b1b837
Signed-off-by: Roy Luo <royluo@google.com>
Signed-off-by: Will McVicker <willmcvicker@google.com>