tty: serial: exynos_tty: stop swallowing tx characters
Assuming count==0, then we still decrement it before skipping the
while-loop. The following test 'if (!count && dma_count)' to determine
if we should use DMA for the remainder will never be entered (as
count!=0 now), thus DMA will never be started for the remainder and we
swallow the characters.

Change the post-decrement to a decrement that only happens if we
actually had any characters to send.

Note 1: The upstream driver is doing the same, see Linux commit [1]
Note 2: We don't actually ever use DMA (no DMA configured in DT), so we
never hit the code path that is being fixed here. I still think this
change is right and we should also follow upstream here.

Link: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=736cd79f483f [1]
      ("serial: samsung: fix DMA for FIFO smaller than cache line size")
Test: TH
Bug: 355413608
Change-Id: Ie80f6e347b86f10fe7c19fdf81f0bec71992082e
Signed-off-by: André Draszik <draszik@google.com>
(cherry picked from commit 198f1b941c624c0f69ccb4b61a3e112931b2aecb)
Signed-off-by: Will McVicker <willmcvicker@google.com>