s2mpg1x-regulator: fix use-after-free: case 1.3.2: for_each_child_of_node() (v6.12 prep)
Case 1.3.2 from below bug:

    for_each_child_of_node() releases nodes during looping using
    of_node_put(), and therefore the node pointer is invalid after the
    loop.

In this case, this is a problem because the driver is using that same
node pointer later as part of regulator_register().

To address this issue, we bump the reference count until
regulator_register() has been called so as to keep the node around past
loop exit.

While at it, we move the test for pdata in s2mpg10_pmic_probe() to
ensure all users can safely dereference it.

Test: TH
Bug: 370679843
Change-Id: Iaadddd75d56085835660d4073d6a2df42d1bd3ea
Signed-off-by: André Draszik <draszik@google.com>
Signed-off-by: Will McVicker <willmcvicker@google.com>