drivers: cpufreq: exynos-acme: fix use-after-free: case 1.3.2: for_each_child_of_node() (v6.12 prep)
Case 1.3.2 from below bug:

    for_each_child_of_node() releases nodes during looping using
    of_node_put(), and therefore the node pointer is invalid after the
    loop.

In this case, this is a problem because the driver is using that same
node pointer later.

To address this issue, we simply bump the reference count so the node
doesn't go out of scope, and add an appropriate devm cleanup handler to
ensure releasing of the node when required.

Test: TH
Bug: 370679843
Change-Id: Ie62cfbdd9b510ffc8547a12ce0b6d2aa2cc408ec
Signed-off-by: André Draszik <draszik@google.com>
Signed-off-by: Will McVicker <willmcvicker@google.com>