mfd: slg51002: use correct value for gpiod_direction_output(_raw) (v6.12 prep)
Linux commit [1] has removed GPIOF_DIR_OUT and GPIOF_DIR_IN.

This causes this OOT Pixel driver to fail compilation.

It also highlights incorrect use of this flag while calling
gpiod_direction_output(). There is no symbolic value to be used as
argument, drivers are meant to pass 0 or 1, as appropriate, see
Documentation/driver-api/gpio/intro.rst in the Linux source tree.

To fix this obvious bug and in preparation of that kernel change,
switch to using the correct value.

Link: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=8c045ca534d0 [1]
      ("gpiolib: legacy: Kill GPIOF_DIR_* definitions")
Test: TH, compiles locally
Bug: 369317190
Change-Id: I3080068b5324c292958b64ad5ff011ca58a8b662
Signed-off-by: André Draszik <draszik@google.com>
Signed-off-by: Will McVicker <willmcvicker@google.com>