pinctrl: slg51002: update to use fwnode instead of of_node (v6.3 prep)
Upstream is removing deprecated OF-specific gpio APIs in [1] and [2] in
particular. As a preparation, we need to update this driver
accordingly.

This commit is similar to commit e642caacd65d ("s2mpg1x-gpio: update to
fwnode instead of of_node") and commit 1c68de2c79c3 ("pinctrl:
slg51000: use device_property_read_u32() for ngpios") in this tree.

Link: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=17bbc46fc9d5 [1]
      ("Merge tag 'gpio-updates-for-v6.3' of git://git.kernel.org/pub/scm/linux/kernel/git/brgl/linux")
Link: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=70d0fc4288da [2]
      ("gpiolib: Get rid of not used of_node member")
Test: TH, compiles locally
Bug: 273738655
Change-Id: I9e5621796ff1d087ba7245ef6f75d2d0dd5e5dcc
Signed-off-by: André Draszik <draszik@google.com>
Signed-off-by: Will McVicker <willmcvicker@google.com>