soc: google: cal-if: fix searching "pmucal_dbg_cpu" node
In pmucal_dbg_init(), the 'node' variable serves two purposes:
    * as a pointer to the "pmucal_dbg" node
    * as a loop iterator for certain node types

Initially, it is set to point to the "pmucal_dbg", and because it is
not released, the subsequent loop will try to find the "pmucal_dbg_cpu"
node below that node.

Judging by the code, 'node' should probably be set to NULL (and
released of course) before entering the loop:
    * the same happens in pmucal_dbg_debugfs_init() ('node' starts as
      NULL)
    * the same happens for all other loops (on loop exit, 'node' will
      be NULL)

I think this makes sense, as only when NULL "pmucal_dbg_cpu" will be
searched for in the full tree.

Test: TH
Bug: 370679843
Change-Id: Ie24d977e80b580dfba359173fe80aa1436071169
Signed-off-by: André Draszik <draszik@google.com>
Signed-off-by: Will McVicker <willmcvicker@google.com>