treewide: plug memleak: case 2.2.2: for_each_child_of_node() (v6.12 prep)
Case 2.2.2 from below bug:

    If returning or breaking early from within
    for_each_child_of_node(), a reference to the current device_node
    remains and needs to be dropped, either right away, or later if
    used past that loop to avoid a memleak.

In all but one of the uses addressed in this commit, np is not needed
past the loop and code can be simplified by switching to
for_each_child_of_node_scoped() which avoids the need for manual
cleanup in early exits from the loop.

The one different use is in exynos-bcm_dbg, where the node is kept
ref'd as part of an early exit from the for_each_child_of_node() loop
and stays around for the lifetime of this driver. In that one case, we
need to add explicit cleanup on error and driver unbind (module
unload).

This commit plugs memleaks in error code paths.

Test: TH
Bug: 370679843
Change-Id: I5dc0333f1e33b94fe0d2e9afb1da3519e6938981
Signed-off-by: André Draszik <draszik@google.com>
Signed-off-by: Will McVicker <willmcvicker@google.com>