pcie: exynos: plug memleak: case 2.2.3: for_each_node_by_type() (v6.12 prep)
Case 2.2.3 from below bug:

    If returning or breaking early from within for_each_node_by_type(),
    a reference to the current device_node remains and needs to be
    dropped, either right away, or later if used past that loop to
    avoid a memleak.

In this case, the np is not needed past the loop. Linux does not have a
scoped version of for_each_node_by_type(), so we simply add a cleanup
handler that will do the necessary unref during early exits from the
loop.

This commit plugs memleaks in error code paths.

Test: TH
Bug: 370679843
Change-Id: Iea686898ba2a01ec3df7e1a57c18b1f8feb4ddcb
Signed-off-by: André Draszik <draszik@google.com>
Signed-off-by: Will McVicker <willmcvicker@google.com>