soc: google: cpupm: shorter spinlock within extern_idle_ip_init()
The current code holds the spin lock for a potentially long time.

In particular this is an issue with an upcoming patch that fixes a
use-after-free case in the of_property_read_string_index() usage in
this loop. In particular that patch will use devm_kmemdup(), which also
acquires a spinlock, which then can cause a locking inversion issue
with other parts of this driver.

Move the spin lock so it protects the critical section only, which will
simplify the use-after-free fix. This makes sense anyway, as spinlock
is meant to be short...

Test: TH
Bug: 374084861
Change-Id: Ica18e28b84a0e3e61cad038afb0f42429c6510fa
Signed-off-by: André Draszik <draszik@google.com>
Signed-off-by: Will McVicker <willmcvicker@google.com>