iommu: samsung: move ->remove_dev_pasid() to blocked_domain (v6.14 prep)
The ->remove_dev_pasid() iommu op was removed upstream in commit
647b7aad1949 ("iommu: Remove the remove_dev_pasid op"). The logic to
attach pasid is now handled through the blocked domain.

This patch implements a basic blocked domain iommu_ops and migrates the
samsung_sysmmu_remove_dev_pasid() callback to the ->set_dev_pasid() in
this blocked domain.

This allows the iommu device to be registered successfully according to
the updated requirements. However, further inspection and refinement of
this implementation is likely necessary.

Bug: 407525102
Test: TH
Change-Id: I4def2337041ef258afff1f556279c2a9a84321e0
Signed-off-by: Carlos Llamas <cmllamas@google.com>