bts: move bw_trip[] to driver
This file currently fails compilation when using a custom
upstream+downstream build environment:

    In file included from .../display/samsung/cal_9845/../exynos_drm_decon.h:30,
                     from .../display/samsung/cal_9845/decon_reg.c:21:
    .../soc/gs/include/soc/google/bts.h:21:27: error: ‘bw_trip’ defined but not used [-Werror=unused-const-variable=]
       21 | static const unsigned int bw_trip[BTS_HIST_BIN - 1] = {
          |                           ^~~~~~~

I'm not sure what's different other than using gcc vs clang (i.e. why
this error isn't seen in the pure downstream build), but declaring the
struct in the header file here certainly means that any code including
this header will get a (useless) copy of the struct.

This is not the intention, the only actual user of the struct is the
bts driver. Move it there.

Test: TH
Fixes: cf5e252f82c9 ("bts: Expose BTS BW metrics")
Bug: 423391571
Change-Id: I422c43242ea12bd5a5d3eb28ed85c7f9155b2f76
Signed-off-by: André Draszik <draszik@google.com>