iommu: samsung: move to ops->domain_alloc_paging()
The ops->domain_alloc() has been deprecated upstream in commit
21c03574df19 ("iommu: Hide ops.domain_alloc behind CONFIG_FSL_PAMU").
Alternative domain-specific options are now in place including the
ops->domain_alloc_paging().

However, this move might be incomplete as the IOMMU_DOMAIN_IDENTITY type
remains unhandled and might require an ops->domain_alloc_identity() to
be setup for this driver. This can be implemented later if this use-case
is really needed though.

Bug: 429438232
Cc: Daniel Mentz <danielmentz@google.com>
Change-Id: I98fa872eaafc8203a0707dffe82277a6b5608964
Signed-off-by: Carlos Llamas <cmllamas@google.com>