vh: clocktrace: Add new clocktrace module
Upstream commit [1] removed the clock_set_rate enable/disable clock
events that were defined in power.h and unused by in-tree code since
2011. However downstream pixel drivers make extensive use of these
clock events.

To maintain the debug and trace support, this commit adds a new
clocktrace module that registers the same clock events removed in [1].

This approach is minimally invasive to the calling drivers, and in
most cases simply requires an update for the header file power.h to
clock.h.

Note: any scripts or tooling enabling the events should be updated
from

/sys/kernel/tracing/events/power/clock_set_rate/
to
/sys/kernel/tracing/events/clock/clock_set_rate/

Link: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=de6fdc076d39 [1]
Bug: 429926222
Change-Id: I31fa3872f7c2b8098c58eb30f9253b81db4487c2
Signed-off-by: Peter Griffin <gpeter@google.com>