mfc: check wait_state even if only crop info is changed
If only the crop information changes, user gets the information
through g_fmt() and g_selection().
At this time, the plane_size information can be initialized
for recalculation, and if the command is sent to the F/W as it is,
it can cause error 57(MFC_REG_ERR_INSUFFICIENT_DPB_SIZE).

Bug: 328149640
Change-Id: I1239985730084d7fe8ecec25d73dc7dc90fc0ac5
Signed-off-by: Ayoung Sim <a.sim@samsung.com>