Compile dtbs with rules_devicetree.
This change uses ruleset from rules_devicetree to build dtb, dtbo and
composite dtb. This decouples dt build rules from kernel_build() as
an isolation.

This change also includes build rules for composite dtbs, which applies
the overlays on base dtbs, ensuring they are applicable. The overlays
(.dtbo) and the composite .dtb files are not included in the dist dir.
However, to ensure the check is executed, a flag file is added.

Bug: 376513172
Change-Id: I9e7797c2b29684c01022158d5eb47f9c7fdfb6f3