iommu/samsung: Track domains on a per-VID basis
The previous commit 90a023c1ba47 ("iommu: samsung: save pointer to iommu
domain at attach_dev/set_dev_pasid time (v6.11 preparation)")
incorrectly assumed a single domain per device. This patch corrects the
logic to track domains on a per-VID basis, ensuring the correct domain
is used, especially in scenarios where multiple VIDs might be in use.

This resolves a potential panic in the fault notifier if called when a
domain for a specific VID is not yet attached or has been detached.
The notifier will now correctly return -EFAULT in such cases, triggering
a panic, as a domain for the specific VID is not attached.

Fixes: 90a023c1ba47 ("iommu: samsung: save pointer to iommu domain at attach_dev/set_dev_pasid time (v6.11 preparation)")
Bug: 407892186
Change-Id: Idc0f76c44a9ac21a464bbf89a664e3e217e549ff
Signed-off-by: Daniel Mentz <danielmentz@google.com>
Signed-off-by: Ryan Huang <tzukui@google.com>