vh: sched: Check task_current_donor() in detach_important_task()
With Proxy Execution enabled, its easy to hit crashes with the
current pixel vendorhooks. I've chased this issue down and it
is happening as the detach_important_task() function calls
set_task_cpu() on a task that is the rq->donor.

So to avoid this, check task_current_donor() on the task
before we consider migrating it.

With this patch, I no longer see the trivial to hit crashes
when booting with sched_proxy_exec=true

Bug: 447263966
Change-Id: I3843485f2b1ce83ca34584c3a31368dc80acd023
Signed-off-by: John Stultz <jstultz@google.com>